helpers.py, create_workbook function
